[#I18] i18n: simplify language toggle impl
Combined with DropdownToggle components since it was the only
instance. Later it can help reimplement the language toggle for better
accessibility.
